get germplasm list contents via germplasm controller
