+++ src/main/java/org/breedinginsight/brapi/v2/services/BrAPIListService.java
@@ -12,6 +12,7 @@ import org.breedinginsight.brapi.v2.dao.BrAPIGermplasmDAO;
 import org.breedinginsight.brapi.v2.dao.BrAPIListDAO;
 import org.breedinginsight.brapps.importer.services.ExternalReferenceSource;
 import org.breedinginsight.model.Program;
+import org.breedinginsight.model.delta.DeltaListDetails;
 import org.breedinginsight.services.exceptions.DoesNotExistException;
 import org.breedinginsight.utilities.Utilities;
 
@@ -19,6 +20,7 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -87,5 +89,13 @@ public class BrAPIListService {
         }
 
         return programLists;
+    }
+
+    public DeltaListDetails getDeltaListDetails(String listDbId, UUID programId) throws ApiException {
+        return listDAO.getDeltaListDetailsByDbId(listDbId, programId);
+    }
+
+    public void deleteBrAPIList(String listDbId, UUID programId, boolean hardDelete) throws ApiException {
+        listDAO.deleteBrAPIList(listDbId, programId, hardDelete);
     }
 }
